pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "live-stream-discord-bot"
version = "1.0.0"
description = "YouTube配信の開始を自動検知してDiscordに通知するPythonアプリケーション"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "live-stream-discord-bot contributors" }
]
keywords = [
    "youtube",
    "discord",
    "webhook",
    "stream",
    "monitoring",
    "notification",
    "vtuber",
    "clean-architecture"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "google-api-python-client>=2.100.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0"
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/live-stream-discord-bot"
Documentation = "https://github.com/YOUR_USERNAME/live-stream-discord-bot#readme"
Repository = "https://github.com/YOUR_USERNAME/live-stream-discord-bot.git"
Issues = "https://github.com/YOUR_USERNAME/live-stream-discord-bot/issues"
Changelog = "https://github.com/YOUR_USERNAME/live-stream-discord-bot/blob/main/CHANGELOG.md"

[project.scripts]
youtube-monitor = "main:main"

[tool.setuptools]
packages = [
    "domain",
    "domain.entities",
    "domain.value_objects",
    "domain.repositories",
    "application",
    "application.use_cases",
    "application.services",
    "application.dto",
    "infrastructure",
    "infrastructure.youtube",
    "infrastructure.discord",
    "infrastructure.persistence",
    "infrastructure.logging",
    "presentation",
    "presentation.cli",
    "config"
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | \.mypy_cache
  | \.pytest_cache
  | __pycache__
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 100
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    ".mypy_cache",
    ".pytest_cache",
    "build",
    "dist"
]
ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "googleapiclient.*",
    "pytest.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov-report=term-missing"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests"
]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "venv/*",
    ".venv/*",
    "*/__pycache__/*",
    "*/site-packages/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod"
]
